fix: Render Conversation action dialogs as singletons from Chat Sidebar level
